test(api): address universal deeppot review comments
